# pkg/shorter.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "shorter.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Shorter"
    }
  ],
  "host": "localhost:11000",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/shorter.Shorter/Create": {
      "post": {
        "operationId": "Shorter_Create",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/shorterCreateLinkResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/shorterCreateLinkRequest"
            }
          }
        ],
        "tags": [
          "Shorter"
        ]
      }
    },
    "/shorter.Shorter/Get": {
      "post": {
        "operationId": "Shorter_Get",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/shorterGetLinkResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/shorterGetLinkRequest"
            }
          }
        ],
        "tags": [
          "Shorter"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "shorterCreateLinkRequest": {
      "type": "object",
      "properties": {
        "link": {
          "type": "string"
        }
      }
    },
    "shorterCreateLinkResponse": {
      "type": "object",
      "properties": {
        "shortLink": {
          "type": "string"
        }
      }
    },
    "shorterGetLinkRequest": {
      "type": "object",
      "properties": {
        "link": {
          "type": "string"
        }
      }
    },
    "shorterGetLinkResponse": {
      "type": "object",
      "properties": {
        "originalLink": {
          "type": "string"
        }
      }
    }
  }
}